Use Normalize.css on all pages via _app.ts
_app.ts is imported as a wrapper for all pages. We don't change its
behaviour, but rely on it being imported to inject the global styles.

This change also adds support to import any arbitrary CSS file as a
side effect, so a transformer has been added to the Jest config to
support imports in tests. It's recommended to continue to use styled-jsx
for stlying components internal to this project for autoprefixing and
style scoping support.