Support using image snapshots to prevent regressions for features
We take a screenshot after every feature test, and compare to the
existing screenshot.

We only run them in Chrome by default to reduce the burden on developers
when creating new tests or updating existing snapshots. This can be
overridden by environment variables.

We resize the images to ensure we always get the same size screenshots,
taking HiDPI displays into account.